Parsing of metadata, details, and init data is basically done.
While we don't hold on to everything that is parsed, we are basically
matching what Heroes.ReplayParser does as far as data it keeps and
roughly the same naming.

While not identical, this library intends to be a faster, more
efficient clone, and so things should be roughly named the same thing.