Big commit: good attribute support.
We can now parse attributes!  Woohoo.  As before, we've tried to include
roughly the same corner cases that Heroes.ReplayParser has about certain
builds having certain quirks.

The biggest change is that we sort of dispensed with the whole notion of
user ID vs working set slot ID vs attribute player ID.  They're clearly
all related, and potentially offset by a constant but otherwise the
same.  This lets us only require one field on the player object and a
simpler search for players by "index", which is what I called the value.

We also implemented our own binary reader, emulating
heroprotocol/Heroes.ReplayParser, instead of trying to use an outside
crate.  The semantics are slightly different in that integers are
big-endian but intended to be read in LSB order, not MSB order, as you
would normally do for big-endian integers.  *shrug*