Add a validate mode of parsing for replays.
This mode goes through some of the basic files in the replay, like the
metadata, init, and details.  This lets us do a quick sanity check on
the replay file to make sure it's actually a replay, and it lets us
generate a UUID based on the replay details.

This mode will serve as a quick way for Stormcaster to make sure a file
looks like a replay, and to get a unique UUID for it so we can avoid
duplicate uploads.